# api/openapi.yaml
openapi: 3.1.0
info:
  title: TiDB DM 可视化平台 API
  version: 0.9.6
  description: |
    后端接口契约 v0.9.6（对齐 DM v7.1.6，见 x-dm-compat；接口形状冻结见 x-freeze）。
    schema 双通道（D11）：`jsonSchema` 为纯标准 JSON Schema 2020-12（属性内无任何 `x-ui-*`），
    全部 UI 元数据在 `formLayout`（`ui` 以 JSON Pointer 为键 + `steps[].groups[].fields[]`）。
    cookieAuth 全局生效；除 /login、/healthz 外 /api/** 未认证 401、无权限 403、首登未改密 40302、CSRF 失败 40303。
    状态码为权威；code=0 仅表示请求已处理；错误统一 data.errors；字段错误码见 x-field-error-codes。
    诊断端点（/task-validate、/task-precheck 各 level、/datasources/test）body 可解析为对象即恒 200 + data.valid；非 JSON/非对象 → 400。
    写侧端点请求体非法 → 422 E_VALIDATION_FAILED；502 仅 DM 控制面不可达 E_DM_UNAVAILABLE。
    带 body 的 state-changing 请求须 Content-Type: application/json；无 body 的 state-changing（如 DELETE、/logout）不强制该头，但仍校验 Origin/CSRF。
    配置落库为物化展开（见 x-config-storage）；读取恒返回有效 per-source 配置。
servers:
  - url: /api
security:
  - cookieAuth: []
paths:
...
  /tasks:
    get:
      operationId: listTasks
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { $ref: "#/components/responses/TaskList" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      operationId: createTask
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: Idempotency-Key
          in: header
          required: true
          description: 非幂等创建必带（端点级定义）；缺失 → 400 E_BAD_REQUEST。
          schema: { type: string }
      requestBody: { $ref: "#/components/requestBodies/TaskWrite" }
      responses:
        "201": { $ref: "#/components/responses/TaskItem" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dm_session
      description: >-
        HttpOnly + SameSite=Lax 会话 cookie（HTTPS 时 Secure）。state-changing 请求
        另需 X-CSRF-Token 双提交：请求头须等于 dm_csrf cookie（非 HttpOnly，Path=/、SameSite=Lax、HTTPS 时 Secure）
        且与服务端会话存储 token 一致；登录/登出轮换 token（防会话固定）。
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: >-
        双提交 CSRF：须等于 dm_csrf cookie 且与服务端会话存储 token 一致（服务端握手，非仅回显比对）；
        缺失/不符 → 403 E_CSRF_FAILED。POST /login 无会话，豁免 CSRF，但仍校验 Origin 与 Content-Type。
      schema: { type: string }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  requestBodies:
...
    TaskWrite:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/TaskWrite" }
...
  responses:
    Error:
      description: 统一错误包（HTTP 状态码为权威；字段级错误在 data.errors）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    TaskList:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Envelope"
              - properties:
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/TaskSummary" } }
                      total: { type: integer }
                      page: { type: integer }
                      pageSize: { type: integer }
    TaskItem:
      description: OK
      headers:
        ETag:
          description: 当前资源 ETag
          schema: { type: string }
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Envelope"
              - properties:
                  data: { $ref: "#/components/schemas/Task" }
...
  schemas:
    Envelope:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer }
        data: { type: [object, "null"] }
        message: { type: string }
      additionalProperties: false
...
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - properties:
            data:
              type: [object, "null"]
              properties:
                errors:
                  type: array
                  items: { $ref: "#/components/schemas/FieldError" }
                riskSummary:
                  type: object
                  properties:
                    error: { type: integer }
                    warning: { type: integer }
    FieldError:
      type: object
      required: [fieldPath, errorCode, message, severity]
      properties:
        fieldPath: { type: string, description: RFC6901 JSON Pointer }
        ruleCode:
          type: string
          enum: [REQUIRED, RANGE, ENUM, REGEX, DEPENDENCY, EXISTENCE, UNIQUE]
        errorCode:
          type: string
          description: >-
            稳定错误码：E_* 见 x-field-error-codes；PRECHECK_* 见 x-precheck-codes。
            两个命名空间互不重叠，独立于信封 code。
          enum:
            - E_FIELD_REQUIRED
            - E_PARAM_RANGE
            - E_PARAM_ENUM
            - E_PARAM_REGEX
            - E_PARAM_DEPENDENCY
            - E_FIELD_EXISTENCE
            - E_FIELD_DUPLICATE
            - PRECHECK_TABLE_CONFLICT
            - PRECHECK_PK_CONFLICT
            - PRECHECK_ROUTE_OVERLAP
            - PRECHECK_TASK_NAME_DUP
            - PRECHECK_CHARSET_TZ_MISMATCH
            - E_SOURCE_UNREACHABLE
            - E_TARGET_AUTH_FAILED
        message: { type: string }
        severity: { type: string, enum: [error, warning, info] }
...
    TaskSummary:
      type: object
      properties:
        name: { type: string }
        state: { type: string, enum: [new, running, paused, stopped, finished, failed] }
        stage: { type: string, description: DM 原生 TaskStage 透传（大写）。 }
        nativeState: { type: string, enum: [Stopped, Running, Finished], description: 与 stage 同值的 DM 原生核对字段。 }
        lag: { type: integer }
        sourceCount: { type: integer }
...
    BlockAllowList:
      type: object
      additionalProperties: false
      required: [schemaPattern, tablePattern]
      properties:
        schemaPattern: { type: string }
        tablePattern: { type: string }
    SourceInstance:
      type: object
      additionalProperties: true
      required: [sourceRef]
      description: >-
        每源覆盖。形状与冻结的 P0 示例 / task-config-schema.md 对齐：routeRules 与 filters 为规则名引用
        （string[]，分别指向 /routes[*].name、/filters[*].name），blockAllowList 为规则对象数组（BlockAllowList）。
        metaSnapshot 为 object-form（binlogName/binlogPos/binlogGtid）；未知字段保留以支持 task.yaml round-trip。
      properties:
        sourceRef: { type: string, minLength: 1 }
        routeRules:
          type: array
          items: { type: string }
          description: 引用 /routes[*].name（规则名，不是内联对象）。
        blockAllowList:
          type: array
          items: { $ref: "#/components/schemas/BlockAllowList" }
        filters:
          type: array
          items: { type: string }
          description: 引用 /filters[*].name（规则名，不是内联对象）。
        metaSnapshot: { type: [object, "null"], additionalProperties: true }
        binlogName: { type: [string, "null"] }
        binlogPos: { type: [integer, "null"] }
        binlogGtid: { type: [string, "null"] }
    TargetDatabase:
      description: >-
        下游 TiDB 目标库（对齐 DM task.yaml 的 target-database，映射见 dm-v7.1.6-field-mapping.md §2）。
        `password` 为纯写入通道（writeOnly: true）：请求可出现，任何读取模型（Task.config 读回、
        /tasks/{name}/yaml 导出、日志/审计）一律不得回显（D16 自有路径）。
      type: object
      additionalProperties: true
      properties:
        host: { type: string }
        port: { type: integer }
        user: { type: string }
        password: { type: string, writeOnly: true }
        security: { type: object, additionalProperties: true }
        session: { type: object, additionalProperties: true }
    Task:
      type: object
      properties:
        name: { type: string }
        rawYaml: { type: string }
        config: { $ref: "#/components/schemas/TaskConfig" }
        sources:
          description: >-
            config.sources 的只读派生投影（同构、同 SourceInstance）；由服务端物化展开后返回，
            不是独立写入入口。写入一律走 TaskConfigWrite/TaskConfigUpdate 的 config。
          type: array
          readOnly: true
          minItems: 1
          items: { $ref: "#/components/schemas/SourceInstance" }
    TaskConfig:
      description: >-
        任务配置结构化模型（写入与读取同构）。多源放 config.sources；未知字段原样保留
        以支持 task.yaml round-trip。落库为物化展开（见 x-config-storage）。
        **任务名 `name` 在写侧顶层（`TaskConfigWrite.name`），不在 `TaskConfig`（config）内**；
        故 `GET /task-schema` 的 `jsonSchema.required` 不得包含 `/name`（唯一真源为顶层 name）。
      type: object
      required: [sources]
      additionalProperties: true
      properties:
        taskMode: { type: string, enum: [all, full, incremental] }
        targetDatabase: { $ref: "#/components/schemas/TargetDatabase" }
        ignoreCheckItems:
          description: >-
            需跳过的预检项名（对齐 DM Task.ignore_checking_items）。取值集合见
            x-dm-compat.precheck.ignoreItemIds；DM 各版本集合不同故不设 enum，
            原样下发，未识别项由 DM 报 20001 ErrConfigCheckItemNotSupport。
          type: array
          items: { type: string }
        sources:
          description: >-
            多源列表（minItems=1）。校验失败恒 422 E_VALIDATION_FAILED，细节只在 data.errors[]：
            字段缺失 → { fieldPath: "/config/sources", errorCode: "E_FIELD_REQUIRED" }；
            空数组 [] → { fieldPath: "/config/sources", errorCode: "E_PARAM_RANGE" }。
            不用 42204 E_PARAM_REQUIRED_ONE（那指"多字段中至少给一个"）；ruleCode 非冻结契约，勿断言。
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/SourceInstance" }
    TaskWrite:
      description: >-
        写入二选一（config 或 rawYaml）。服务端须显式区分：**同给** → 422 `E_PARAM_MUTUALLY_EXCLUSIVE`(42203)；
        **都不给** → 422 `E_PARAM_REQUIRED_ONE`(42204)；其余写侧 schema 违规 → 422 `E_VALIDATION_FAILED`(42201)。
        JSON Schema 的 oneOf 分不出「都给/都缺」，故不依赖 oneOf 自动判定。
      oneOf:
        - $ref: "#/components/schemas/TaskConfigWrite"
        - $ref: "#/components/schemas/TaskYamlWrite"
    TaskConfigWrite:
      type: object
      required: [name, config]
      additionalProperties: false
      properties:
        name: { type: string, pattern: '^[a-zA-Z0-9_-]{1,64}$' }
        config: { $ref: "#/components/schemas/TaskConfig" }
    TaskYamlWrite:
      type: object
      required: [name, rawYaml]
      additionalProperties: false
      properties:
        name: { type: string, pattern: '^[a-zA-Z0-9_-]{1,64}$' }
        rawYaml: { type: string }
...